pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "photosort"
version = "2.0.0"
requires-python = ">=3.10"
dependencies = [
    "Pillow>=10.0",
    "customtkinter>=5.2",
    "hachoir>=3.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
photosort = "photosort.cli:main"
photosort-gui = "photosort.gui.app:main"

[tool.setuptools.packages.find]
where = ["src"]
